[TASK] Move and document vital first-sequence configuration to own file
This takes the content rendering templates storage array and moves it to an AdditionalConfiguration.php file included in the Build directory. This file must be copied to typo3conf in order to avoid possible "no rendering definition" errors.

A fallback is introduced, adding the storage array if it is not set - but at the trade-off that *some* "no rendering definition" errors can occur when *others* will not - better than nothing.