[FEATURE] Menutype optional Properties Partial rendering
Adds an optional method to quickly define new fields according to menu type, in a way similar to how the Partials for each menu type are rendered and using the same naming scheme for consistency.

Example:

```
Resources/Private/Partials/Properties/Menu/RecentlyUpdated.html
Resources/Private/Partials/Properties/Menu/SiteMap.html
```

Inside file:

```
<flux:form.sheet name="menu" label="LLL:sheets.menu">
	<flux:field.select name="settings.menu.limit" items="0,5,10,15,25,50" label="LLL:fields.menuLimit" />
</flux:form.sheet>
```

No `f:section` is needed inside the Properties Partial and the template itself is optional - if it exists, it is used, if not, it is ignored.