[BUGFIX] Reference correct variable in NoHeader layout
The NoHeader.html layout referenced to the container tag name using content.settings.container.types. This variable is empty since the container tag assignments are not part of the settings of an individual content element. settings.container.types is the right reference and also how the Default.html layout does it already.