[TASK] Render images using vhs/imgResource and uid
fluid's ImageViewHelper renders images with extbase ImageService, but vhs uses imgResource. imgResource can be easily hooked/extended to provide advanced processing of images (example: EXT:tkcropthumbs), whereas this is currently not as easy for ImageService. Using imgResource is thus needed for compatibility with extensions like tkcropthumbs. In the general case, nothing should change by using the vhs ViewHelper.

Using the image UID instead of the direct URL is needed for the ViewHelper to fetch the sys_file_reference information of the image which can hold further processing information. This would be impossible if only the corresponding image URL is provided.