[BUGFIX] Fix shortcut type
By default, the "records" argument is saved without a prefix as the default
table is "tt_content". The previous fix will discard all records in this case.
It will also leave empty elements in contentUids for each element it filters out
because array_filter was not called on the array_map result.

Also, the test did not actually test anything because of a wrong "method()" call.