[TASK] Fetch raw record before rendering provider field
Because I was unable to explain the case for introducing backwards compatibility instead of breaking in https://review.typo3.org/#/c/43408/ this change is now necessary. Others may be required later.

It is the hope of this author that "uid" will still considered a single value field even if "CType" is not.